:truck: [#200] Move setup_config files to setup_configuration directory
