+++ src/nrc/tests/setup_configuration/files/setup_config_auth_config.yaml
